change fields

+++ App/models.py
@@ -7,5 +7,5 @@ class city(models.Model):
 class Component(models.Model):
     id = models.AutoField(primary_key=True,db_column="Id")
     name = models.CharField(max_length=255,null=False,blank=False,db_column="Name")
-    phone = models.IntegerField(null=False,blank=False,db_column="Phone_Number")
+    phone = models.BigIntegerField(null=False,blank=False,db_column="Phone_Number")
     city = models.ManyToManyField(city)
